backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
tavily-python
chromadb
reportlab
pydantic
python-dotenv
aiohttp
